electron/module1/server/requirements.txt
fastapi
uvicorn
transformers
torch
sentence-transformers
chromadb
pytesseract
pillow
python-docx
pymupdf
langdetect
google-generativeai
pydantic
python-dotenv
sentencepiece
python-multipart
spacy
markdown2
reportlab
